A bit of cleanup to build under windows.
Change a couple function definitions in datatype.h from extern
to OMPI_DECLSPEC since they are never defined anywhere else, and
need to be for the build.

Added the datatype.h file to ptl_self so it can see the inline
convertor_pack routines in that file when not in Debug mode.