Windows build checkpoint.
Many of the changes are include cleanups, and the
addition of pls, ras, and rds to libmpi for the tools
directory.

Some of the functionality is currently stubbed out.
Need to return to implement things properly so we
actually run under Windows. But I am focusing on just
building at the moment :)