Get rid of the storage in the convertor. It wasn't working as expected in all the cases. Instead replace it with a better solution, which work even for fragments received not in order. However, this solution work only on the current supported modes in ompi (homogeneous & heterogeneous with endianess).
The method is tricky. We will rely on 2 partial unpacks. First we will find
a byte that is not on the data to unpack, and we will pad the data with this
byte. Once we have the full length as expected, we will unpack the data, and
all the bytes in the unpacked form which do not match the unused byte will be
copied into the user buffer. This way we will reconstruct the unpacked data
in 2 times, once for the begining and once for the end.