Move r22593, r22594, r22604, r22607, r22612, r22649, r22662, r22670 and r22686.
Check whether file is opened on network file-system. If file does not exist, check the directory it lives in... Maybe used by caller, trying to open mmap() on NFS, Lustre or Panasas (thanks Sam). For now, this is used to warn about the usage of mmap on such FS.

This commit closes #2256.