The patches looks good. I'll move them directly to the 1.5 per RM request.
Fixes #2335: Move IB locking fixes to v1.5 (these are in 1.4.2)

r22676
This fixes a bug where sometimes the rcache lock would be
dropped when it wasn't actually held.
Also includes some minor copytight header additions that
were missed in previous checkins

r22678
Fixes bug where the wrong lock is taken in mca_btl_openib_alloc
when protecting the no_wqe_pending_frags list.

r22679
Destroying an rcache vma while the rcache lock is held
as this can result in a low level free of memory which
can require the rcache lock resulting in a deadlock

r22682
The ib_procs list in the openib btl is accessed without
the ib lock in some cases. This causes races when running
multithreaded. This patch adds protection of the ib_procs
list with the ib_lock.


r22683
The release of memory used by registration lists in rcaches
must be delayed until the rcache lock is not held or
deadlock can occur.
Should not deregister memory with the rcache lock held
otherwise a deadlock can occur as the lower level infiniband
libraries can free memory.

r22684
This fixes #2265 removing a race in the openib btl endpoint when
increasing sequence numbers.