Fixes #2482: Fix bug in --output-filename
r23380
The --output-filename option currently mixes the output from all processes with the same rank, but different jobids. Thus, the output from comm_spawn'd processes gets intermingled with their parents and each other.

Adding the jobid to the output filename solves the problem.

Thanks to Jody for pointing this out!