Fixes #2669.
Apparently, gcc 4.4.x and 4.5.x complain about the ''possibility'' of
us calling free() on a non-heap variable.  We know that this case can
never happen because the refcount will absolutely not go to zero
here.  We think it may be gcc being a bit too aggressive on the
warnings.

However, since this happens with gcc 4.4.x and 4.5.x, and since gcc
4.5.x ship in RHEL6 and Fedora 14 (and others), someone '''will'''
complain about this in the future, so we might as well code around it
so that we don't have to keep explaining "despite the warning, it's
really ok."

The workaround is pretty simple: just OBJ_RELEASE the values from
ompi_mpi_comm_parent before it is re-assigned to the new
intercommunicator.  Then the compiler's static code analysis can't
possibly tell that it's not a heap variable, and we're ok.

So yes, we are still calling OBJ_RELEASE on a non-heap variable.  But
free() '''will never be called''' on it because of the refcount.