This is a hand-made patch by Ralph based on the following list of commits:
Fixes #2625: Fix oversubscribe flag

r24037
Properly count local procs to set oversubscribed condition

r23126
Remove the "linux" paffinity component (i.e., the one that was based
on the now-defunct PLPA) -- the new hwloc component supersedes it.

So long, PLPA -- we loved ya!

r24127
Not totally necessary, but good form - init the oversubscribed field in the orte_nid_t object

r24130
One more time, into the breach!

Restore the use of override_oversubscribe to indicate that the data source for resources on the backend nodes used in mapping is unreliable. In this situation (e.g., data came from hostfile, or we are just using localhost because nothing was provided), we don't trust the oversubscribe condition passed by the mapper. Instead, we check locally to ensure we set sched_yield correctly.