Fixes #2671: Move MPI_Init_thread C/R fix to v1.5 branch
r24194
MPI_Init_thread is used in place of MPI_Init, so for the checkpoint/restart functionality it must correctly init the C/R functionality instead of simply making a critical section. This allows the C/R thread to be started properly.

Thanks to Takayuki Seki for finding this bug.