Fixes #2962: Fix windows build for v1.5 branch
---svn-pre-commit-ignore-below---
r25724
Add an option for enabling/disabling the compatible support for windows xp. For example, the InterlockedCompareExchange64 support, which doesn't exist on XP. Enabling this option can make sure that the binary built on a Windows 7/Vista could also be run on Windows XP correctly.

r25725
Disable the debugger support for Windows.

r25726
Do not define this if we want Win XP compatible build.